+++ app/src/main/AndroidManifest.xml
@@ -30,6 +30,10 @@
         <activity android:name=".ui.AmbLogActivity"/>
         <activity android:name=".ui.ShakeActivity"/>
         <activity android:name=".ui.ScheduleActivity"/>
+        <activity android:name=".ui.ScheduleViewActivity"/>
+        <activity android:name=".ui.PhotoActivity"/>
+        <activity android:name=".ui.DrawActivity"/>
+        <activity android:name=".ui.AnimationActivity"/>
 
         <receiver android:name=".Broadcast.MyBroadcastReceiver">
             <intent-filter>
@@ -46,6 +50,7 @@
             />
 
         <service android:name=".Service.ScheduleService"/>
+        <service android:name=".Service.BgMusicService" android:exported="true" android:process=":remote"/>
 
     </application>
