+++ app/src/main/java/com/lz/www/ambrm/MainActivity.java
@@ -1,5 +1,7 @@
 package com.lz.www.ambrm;
 
+import android.app.AlarmManager;
+import android.app.PendingIntent;
 import android.content.ContentResolver;
 import android.content.ContentValues;
 import android.content.Intent;
@@ -13,10 +15,14 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.lz.www.ambrm.Service.BgMusicService;
 import com.lz.www.ambrm.model.Contacts;
 import com.lz.www.ambrm.model.HttpResponse;
 import com.lz.www.ambrm.net.HttpService;
 import com.lz.www.ambrm.ui.HomeActivity;
+import com.lz.www.ambrm.ui.ScheduleActivity;
+import com.lz.www.ambrm.ui.ScheduleViewActivity;
+import com.lz.www.ambrm.util.CommonUtils;
 import com.lz.www.ambrm.util.Config;
 import com.lz.www.ambrm.util.JsonUtils;
 import com.lz.www.ambrm.util.SPUtils;
@@ -56,8 +62,11 @@ public class MainActivity extends AppCompatActivity {
 //                resolver.insert(Config.AMBLOG_URI, values);
 //                Toast.makeText(MainActivity.this,"添加成功",Toast.LENGTH_SHORT).show();
 
-                SPUtils.put(MainActivity.this,"test","我是测试");
+                //SPUtils.put(MainActivity.this,"test","我是测试");
 
+                //启动service
+                Intent intent=new Intent(MainActivity.this,BgMusicService.class);
+                startService(intent);
 
             }
         });
@@ -84,10 +93,33 @@ public class MainActivity extends AppCompatActivity {
 //                    textView.setText(result);
 //                }
 
-               Object time = SPUtils.get(MainActivity.this,"ScheduleTime","0");
-                Object content=SPUtils.get(MainActivity.this,"ScheduleContent","0");
-                Toast.makeText(MainActivity.this,time.toString()+" "+content.toString(),Toast.LENGTH_SHORT).show();
+//               Object time = SPUtils.get(MainActivity.this,"ScheduleTime","0");
+//                Object content=SPUtils.get(MainActivity.this,"ScheduleContent","0");
+//                Toast.makeText(MainActivity.this,time.toString()+" "+content.toString(),Toast.LENGTH_SHORT).show();
 
+//                //定时提醒
+//                Intent it=new Intent(MainActivity.this,ScheduleViewActivity.class);
+//                it.putExtra("content","hahahaha");
+//
+//                PendingIntent pi=PendingIntent.getActivity(MainActivity.this,0,it,0);
+//
+//                AlarmManager alarmManager=(AlarmManager)getSystemService(ALARM_SERVICE);
+//
+//                long sTimes=0;
+//                try {
+//                    Date sTime= CommonUtils.ConverToDate("2016-7-11 16:22:00");
+//                    sTimes=sTime.getTime();
+//                } catch (Exception e) {
+//                    e.printStackTrace();
+//                }
+//                alarmManager.set(AlarmManager.RTC_WAKEUP,sTimes,pi);
+//
+//                Toast.makeText(MainActivity.this,"定义闹钟完成",Toast.LENGTH_SHORT).show();
+
+
+                //停止service
+                Intent intent=new Intent(MainActivity.this,BgMusicService.class);
+                stopService(intent);
             }
         });
