+++ app/src/main/java/com/lz/www/ambrm/Service/BgMusicService.java
@@ -1,0 +1,76 @@
+package com.lz.www.ambrm.Service;
+
+import android.app.Service;
+import android.content.Intent;
+import android.media.MediaPlayer;
+import android.os.IBinder;
+import android.support.annotation.Nullable;
+
+import com.lz.www.ambrm.R;
+
+import java.io.IOException;
+import java.util.IllegalFormatException;
+
+/**
+ * Created by Administrator on 2016-07-12.
+ */
+public class BgMusicService extends Service {
+
+    private MediaPlayer mediaPlayer;
+
+    @Nullable
+    @Override
+    public IBinder onBind(Intent intent) {
+        return null;
+    }
+
+    @Override
+    public void onCreate() {
+        try {
+            mediaPlayer=new MediaPlayer();
+            mediaPlayer=MediaPlayer.create(BgMusicService.this, R.raw.dxj );
+            mediaPlayer.prepare();
+        }catch (IllegalStateException ex){
+            ex.printStackTrace();
+        }catch (IOException ex) {
+            ex.printStackTrace();
+        }
+        super.onCreate();
+    }
+
+
+    @Override
+    public int onStartCommand(Intent intent, int flags, int startId) {
+        mediaPlayer.start();
+        mediaPlayer.setOnCompletionListener(new MediaPlayer.OnCompletionListener() {
+            @Override
+            public void onCompletion(MediaPlayer mediaPlayer) {
+                try {
+                    mediaPlayer.start();//循环播放
+                }catch (IllegalStateException e){
+                    e.printStackTrace();
+                }
+            }
+        });
+
+        mediaPlayer.setOnErrorListener(new MediaPlayer.OnErrorListener() {
+            @Override
+            public boolean onError(MediaPlayer mediaPlayer, int i, int i1) {
+                try {
+                    mediaPlayer.release();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+                return false;
+            }
+        });
+        return super.onStartCommand(intent, flags, startId);
+    }
+
+    @Override
+    public void onDestroy() {
+        mediaPlayer.stop();
+        mediaPlayer.release();
+        super.onDestroy();
+    }
+}
