+++ app/src/main/java/com/lz/www/ambrm/model/Point.java
@@ -1,0 +1,23 @@
+package com.lz.www.ambrm.model;
+
+/**
+ * Created by Administrator on 2016-07-16.
+ */
+public class Point {
+    private float x;
+
+    private float y;
+
+    public Point(float x, float y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public float getX() {
+        return x;
+    }
+
+    public float getY() {
+        return y;
+    }
+}
