+++ app/src/main/java/com/lz/www/ambrm/ui/Fragment/SbuFragment.java
@@ -1,0 +1,24 @@
+package com.lz.www.ambrm.ui.Fragment;
+
+
+import android.app.Fragment;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.lz.www.ambrm.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class SbuFragment extends Fragment {
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View view=inflater.inflate(R.layout.fragment_fa_sbu,container,false);
+        return view;
+    }
+}
