--- app/src/main/java/com/lz/www/ambrm/ui/MeFragment.java
@@ -1,7 +1,0 @@
-package com.lz.www.ambrm.ui;
-
-/**
- * Created by Administrator on 2016/6/10.
- */
-public class MeFragment {
-}
