+++ app/src/main/res/layout/activity_photo.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    >
+
+    <ImageView
+        android:id="@+id/imgBigPhoto"
+        android:background="#000000"
+        android:layout_margin="5dp"
+        android:layout_weight="1"
+        android:src="@drawable/mn1"
+        android:layout_width="match_parent"
+        android:layout_height="0dp" />
+
+
+    <com.lz.www.ambrm.ui.MyRecyclerView
+        android:id="@+id/rvPhoto"
+        android:layout_marginTop="2dp"
+        android:layout_gravity="center_vertical"
+        android:background="#FF0000"
+        android:layout_width="match_parent"
+        android:layout_height="120dp"/>
+
+</LinearLayout>
