+++ app/src/main/res/layout/activity_schedule.xml
@@ -30,7 +30,8 @@
 
     <EditText
         android:id="@+id/etSchedule"
+        android:drawableLeft="@drawable/b3"
         android:layout_width="match_parent"
-        android:layout_height="50dp" />
+        android:layout_height="100dp" />
 
 </LinearLayout>
