+++ app/src/main/res/layout/fragment_dao.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.lz.www.ambrm.ui.DaoFragment">
+    tools:context=".ui.Fragment.DaoFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <ListView
