+++ app/src/main/res/layout/fragment_shu.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SbuFragment"
+    tools:context=".ui.Fragment.SbuFragment"
     android:background="#00ff00">
 
     <!-- TODO: Update blank fragment layout -->
